requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
piexif>=1.1.3
jinja2>=3.1.0
pydantic-ai>=0.0.23
openai>=1.0.0
plotly>=5.18.0
pandas>=2.0.0
-e . 